… Source/Lib/Common/Codec/EbPictureBufferDesc.h
… Source/Lib/Common/Codec/EbPictureControlSet.c
… Source/Lib/Common/Codec/EbPictureDecisionQueue.c
… Source/Lib/Common/Codec/EbPictureManagerQueue.c
… Source/Lib/Common/Codec/EbSystemResourceManager.c
… Source/Lib/Encoder/Codec/EbEncHandle.c