There are too many malloc times when initiate SvtAv1EncApp. It makes initiation too slow on server and this can be relieved by reduce malloc times in block way.

Vtune show that `mode_decision_context_ctor` takes 30.4% cpu time of eb_init_encoder's 83.5%; and `eb_object_wrapper_ctor` takes 28.2% cpu time of eb_init_encoder's 83.5%. I tried block way to reduce malloc times in this two functions.

Tested by:
```
SvtAv1EncApp -enc-mode 8 -w 1920 -h 1080 -fps 60 irefresh-type 2  -i 1920x1080.yuv 
-q 32 -b output.bin -bit-depth 8 -stat-report  -n 300 
```
Here are some result comparation:

|        | build_type | init_time | malloc_times |
| ------ | ---------- | --------- | ------------ |
| after  | Debug      | 6.4660s   | 331625       |
|        | Release    | 5.6350s   | 331625       |
| before | Debug      | 7.9340s   | 1964126      |
|        | Release    | 6.6300s   | 1964126      |

```
-------------------------------------------                      
SVT-AV1 Encoder                                                  
SVT [version]:  SVT-AV1 Encoder Lib v0.7.5                       
SVT [build]  :  GCC 7.4.0        64 bit                          
LIB Build date: Dec 13 2019 16:09:22                             
-------------------------------------------                      
Number of logical cores available: 112                           
Number of PPCS 159                                               
-------------------------------------------                      
SVT [config]: Main Profile      Tier (auto)     Level (auto)     
SVT [config]: EncoderMode                                       : 8                                                               
SVT [config]: EncoderBitDepth / EncoderColorFormat / CompressedTenBitFormat      : 8 / 1 / 0                                      
SVT [config]: SourceWidth / SourceHeight                        : 1920 / 1080                                                     
SVT [config]: FrameRate / Gop Size                              : 60 / 64                                                         
SVT [config]: HierarchicalLevels / BaseLayerSwitchMode / PredStructure           : 4 / 0 / 2                                      
SVT [config]: BRC Mode / QP  / LookaheadDistance / SceneChange  : CQP / 32 / 33 / 0            
SVT Memory Usage:                                                
    total allocated memory:       35.18 GB 1964126 times         
        malloced memory:          24.15 GB 1346516 times         
        callocated memory:        1.39 GB 266584 times           
        allocated aligned memory: 9.64 GB 351026 times           
    mutex count: 2865                                            
    semaphore count: 2008                                        
    thread count: 536                                            
    hash table fulless: 0.469574, hash bucket is too full
-----------------------------------------------------------
After optimization:
SVT Memory Usage:                                                
    total allocated memory:       35.39 GB 331625 times          
        malloced memory:          24.31 GB 125424 times          
        callocated memory:        1.39 GB 178961 times           
        allocated aligned memory: 9.69 GB 27240 times            
    mutex count: 2865                                            
    semaphore count: 2008                                        
    thread count: 536                                            
    hash table fulless: 0.080355, hash bucket is healthy
```