@zhonghongcheng 2019-12-17
> > It seems that test failed on macOS clang build:
> > ```
> > /Users/travis/.travis/functions: line 109:  7128 Segmentation fault: 11  SvtAv1EncApp -enc-mode 8 -i akiyo_cif.y4m -n 150 -b test1.ivf
> > The command "SvtAv1EncApp -enc-mode 8 -i akiyo_cif.y4m -n 150 -b test1.ivf" exited with 139.
> > ```
> > 
> > 
> > I do not know how to fix this. Can anyone help to tell me what to do?
> > I am a rookie and this is my first PR:)
> 
> You need to reproduce this error and fix the bug. Do you have access to a macOS machine?

The problem is that I do not have access to a macOC machine...