@zhonghongcheng 2019-12-19
I tested on `macOS 10.15.2 Catalina with Apple LLVM 11.0.0 (clang-1100.0.33.8)` and `macOS 10.14.4 Mojave with Apple LLVM 10.0.1 (clang 1001.0.46.4)`, and both passed. Can not reproduce the error in CI. What should I do then?