@zhonghongcheng 2019-12-19
Thanks @1480c1 ! It seems macOS compile is alright after rebase on master.