@zhonghongcheng 2019-12-19
I would like to ask that in function `EbEncHandle.c: allocate_frame_buffer`. There is a hardcoded line [`input_picture_buffer_desc_init_data.buffer_enable_mask = PICTURE_BUFFER_DESC_FULL_MASK;`](https://github.com/OpenVisualCloud/SVT-AV1/blob/master/Source/Lib/Encoder/Codec/EbEncHandle.c#L3691) , which makes options on `buffer_enable_mask` above useless. I am confused about this.